# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AutoSSL 证书管理API文档",
        "title": "AutoSSL API",
        "termsOfService": "http://example.com/terms",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/apiv1",
    "paths": {
...
        "/dashboard/count-stats": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取当前用户的总证书数、access数量、服务组数量、DNS供应商数量",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "仪表板"
                ],
                "summary": "获取用户数量统计",
                "responses": {
                    "200": {
                        "description": "数量统计数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dashboard.CountStats"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "获取统计数据失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/dashboard/pending-workflows": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取最近5条待开始工作流，包含证书名、部署类型、创建时间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "仪表板"
                ],
                "summary": "获取待开始工作流",
                "responses": {
                    "200": {
                        "description": "待开始工作流列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dashboard.PendingWorkflow"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "获取待开始工作流失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/dashboard/recent-deployments": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "获取最近5条部署状态，包含证书名、部署类型、更新时间、状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "仪表板"
                ],
                "summary": "获取最近部署状态",
                "responses": {
                    "200": {
                        "description": "最近部署列表",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dashboard.DeploymentInfo"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "获取部署状态失败",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dashboard.CountStats": {
            "type": "object",
            "properties": {
                "total_accesses": {
                    "type": "integer"
                },
                "total_certificates": {
                    "type": "integer"
                },
                "total_dns_providers": {
                    "type": "integer"
                },
                "total_server_groups": {
                    "type": "integer"
                }
            }
        },
        "dashboard.DeploymentInfo": {
            "type": "object",
            "properties": {
                "certificate_name": {
                    "type": "string"
                },
                "deployment_type": {
                    "description": "\"server_group\" 或 \"access\"",
                    "type": "string"
                },
                "last_updated": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "target_name": {
                    "type": "string"
                }
            }
        },
        "dashboard.PendingWorkflow": {
            "type": "object",
            "properties": {
                "certificate_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deployment_type": {
                    "description": "\"server_group\" 或 \"access\"",
                    "type": "string"
                },
                "target_name": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}